backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
alembic==1.13.1
asyncpg==0.29.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
websockets==12.0
pydantic-settings==2.1.0
psycopg2-binary==2.9.9
email-validator==2.1.0.post1
bcrypt==3.2.2
pandas
openpyxl
reportlab
werkzeug
redis==5.0.1
arq==0.25.0
pymupdf==1.26.7
Pillow
